+++ src/services/form-validation/validateSubmission.ts
@@ -16,7 +16,10 @@ import {
 } from './validators'
 import generateCivicaNameRecordElements from '../generateCivicaNameRecordElements'
 import generateFreshdeskDependentFieldElements from '../generateFreshdeskDependentFieldElements'
-import { parseDateValue } from '../generate-default-data'
+import {
+  parseDateValue,
+  DATE_ELEMENT_SUBMISSION_MODEL_FORMAT,
+} from '../generate-default-data'
 import { localisationService } from '@oneblink/apps'
 
 export const RECAPTCHA_OFFLINE_MESSAGE =
@@ -373,12 +376,12 @@ export default function validateSubmission({
               dateOnly: true,
               format: (date: Date) => localisationService.formatDate(date),
               earliest: parseDateValue({
-                dateOnly: true,
+                dateFormat: DATE_ELEMENT_SUBMISSION_MODEL_FORMAT,
                 daysOffset: fromDateDaysOffset,
                 value: fromDate,
               }),
               latest: parseDateValue({
-                dateOnly: true,
+                dateFormat: DATE_ELEMENT_SUBMISSION_MODEL_FORMAT,
                 daysOffset: toDateDaysOffset,
                 value: toDate,
               }),
@@ -427,12 +430,12 @@ export default function validateSubmission({
             ...validators.datetime(value, {
               format: (v: Date) => localisationService.formatDatetime(v),
               earliest: parseDateValue({
-                dateOnly: false,
+                dateFormat: undefined,
                 daysOffset: fromDateDaysOffset,
                 value: fromDate,
               }),
               latest: parseDateValue({
-                dateOnly: false,
+                dateFormat: undefined,
                 daysOffset: toDateDaysOffset,
                 value: toDate,
               }),