+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- allow screen readers to ignore optionSet images
+
 ## [7.2.0] - 2025-01-29
 
 ### Added